% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maltese.R
\docType{package}
\name{maltese-package}
\alias{maltese-package}
\alias{maltese}
\title{maltese: Machine Learning for Time Series}
\description{
This package facilitates using machine learning algorithms such
  as neural networks and deep learning to make forecasts.
}
