% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{mlts_transform}
\alias{mlts_transform}
\title{Transform a time series to a machine learning-friendly format}
\usage{
mlts_transform(dt, y, p = 1, xreg = NULL, granularity = NULL,
  extrasAsFactors = FALSE, start = c("Mon", "Sun"))
}
\arguments{
\item{dt}{Dates (or date-times) corresponding to time series \code{y}}

\item{y}{an evenly spaced time series as a numeric vector or factor}

\item{p}{Number of previous observations to turn into features (think AR(p))}

\item{xreg}{Optionally, a vector or matrix of external regressors, which
must have the same number of rows as y}

\item{granularity}{One of: seconds, minutes, hours, days, weeks, months,
quarters, years. If not specified, will attempt to detect.}

\item{extrasAsFactors}{Whether to output extra features as factors or
numeric (default). If TRUE, some (like day of week or month) will be
ordered factors.}

\item{start}{Which day does the week start? "Sun"day or "Mon"day (default)?}
}
\value{
A \code{data.frame} suitable for supervised learners with columns:
  \describe{
    \item{dt}{The date or date-time (same as \code{dt})}
    \item{y}{The time series}
    \item{mlts_lag_k}{The previous k-th observation}
    \item{mlts_extras_?}{Extra features like hour of day, day of the week, month of the year, etc.}
  }
}
\description{
Performs the necessary transformations to make a univariate
  time series acceptable for use with machine learning models like neural
  networks.
}
\examples{
data("r_enwiki", package = "maltese")
mlts <- mlts_transform(head(r_enwiki$date), head(r_enwiki$pageviews))
}
