% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{mlts_transform}
\alias{mlts_transform}
\title{Transform a time series to a machine learning-friendly format}
\usage{
mlts_transform(.data, .dt, .y, p = 1, xreg = NULL, granularity = NULL,
  extras = FALSE, extrasAsFactors = FALSE, start = c("Mon", "Sun"))
}
\arguments{
\item{.data}{A tidy data.frame/tbl}

\item{.dt}{Name of the column in \code{.data} containing the date-times;
does not need to be a character}

\item{.y}{Name of the column in \code{.data} containing an evenly spaced
time series as a numeric vector or factor; does not need to be a character}

\item{p}{Number of previous observations to turn into features (think AR(p))}

\item{xreg}{A character vector of column names of external regressors in \code{.data}}

\item{granularity}{One of: seconds, minutes, hours, days, weeks, months,
quarters, years. If not specified, will attempt to detect.}

\item{extras}{Whether maltese will create new features (like day of week)}

\item{extrasAsFactors}{Whether to output extra features as factors or
numeric (default). If TRUE, some (like day of week or month) will be
ordered factors.}

\item{start}{Which day does the week start? "Sun"day or "Mon"day (default)?}
}
\value{
A \code{data.frame} suitable for supervised learners with columns:
  \describe{
    \item{dt}{The date or date-time (same as \code{dt})}
    \item{y}{The time series}
    \item{mlts_lag_k}{The previous k-th observation}
    \item{mlts_extras_?}{Extra features like hour of day, day of the week, month of the year, etc.}
  }
}
\description{
Performs the necessary transformations to make a univariate
  time series acceptable for use with machine learning models like neural
  networks.
}
\examples{
data("r_enwiki", package = "maltese")
mlts <- mlts_transform(head(r_enwiki), date, pageviews)
}
